chore: break out search modal from GlobalSearchButton
